refactor: core/material/stress_states.py を bases.py へ改名し命名衝突を解消
core/stress_state.py (StressState dataclass) との単数/複数の紛らわしさを除去。
外部 API (from manforge.core.material import ...) は __init__.py の re-export で維持。
