+++ src/manforge/core/material/__init__.py
@@ -2,11 +2,11 @@
 
 Re-exports MaterialModel and its three stress-state base classes so
 that ``from manforge.core.material import MaterialModel3D`` continues to
-work after the split into ``base.py`` / ``stress_states.py``.
+work after the split into ``base.py`` / ``bases.py``.
 """
 
 from manforge.core.material.base import MaterialModel
-from manforge.core.material.stress_states import (
+from manforge.core.material.bases import (
     MaterialModel3D,
     MaterialModelPS,
     MaterialModel1D,
