+++ src/manforge/core/material/bases.py
